simplify transition type
